Russett data

The Russett data set (Russett, 1964) is studied in Gifi (1990). Three blocks of variables have been defined for 47 countries. The first block is related to "Agricultural Inequality", the second to "Industrial Development", and the last one describes the "Political Instability". Russett collected this data to study relationships between Agricultural Inequality, Industrial Development and Political Instability. Russett's hypotheses can be formulated as follows: It is difficult for a country to escape dictatorship when its agricultural inequality is above-average and its industrial development below-average.

X1

Agricultural Inequality

GINI: Inequality of land distribution,

FARM: Percentage of farmers that own half of the land,

RENT: Percentage of farmers that rent all their land.

X2

Industrial Development

GNPR: Gross national product per capita ($1955),

LABO: Percentage of labor forced employed in agriculture.

X3

Political Instability

INST: Instability of executive (45-61),

ECKS: Number of violent internal war incidents (46-61),

DEAT: Number of people killed as a result of civic group violence (50-62),

DEMOSTAB: Stable democracy,

DEMOINST: Unstable democracy,

DICTATOR: Dictatorship.

Usage

data(Russett)

Format

A data frame with 47 rows and 12 variables.

References

Russett B.M. (1964), Inequality and Instability: The Relation of Land Tenure to Politics, World Politics 16:3, 442-454.

Gifi, A. (1990), Nonlinear multivariate analysis, Chichester: Wiley.

Examples

#Loading of the Russett dataset
data(Russett)
#Russett is partitioned into three blocks (X_agric, X_ind, X_polit)
X_agric <- Russett[, c("gini", "farm", "rent")]
X_ind <- Russett[, c("gnpr", "labo")]
X_polit <- Russett[, c("inst", "ecks", "death", "demostab",
                       "demoinst", "dictator")]
A <- list(X_agric, X_ind, X_polit)