Available methods for RGCCA

List the methods that can be used with the rgcca function.

Usage

available_methods()

Value

A vector of the methods implemented with the rgcca function.

Examples

available_methods()
#>  [1] "rgcca"     "sgcca"     "pca"       "spca"      "pls"       "spls"     
#>  [7] "cca"       "ifa"       "ra"        "gcca"      "maxvar"    "maxvar-b" 
#> [13] "maxvar-a"  "mfa"       "mcia"      "mcoa"      "cpca-1"    "cpca-2"   
#> [19] "cpca-4"    "hpca"      "maxbet-b"  "maxbet"    "maxdiff-b" "maxdiff"  
#> [25] "sabscor"   "ssqcor"    "ssqcov-1"  "ssqcov-2"  "ssqcov"    "sumcor"   
#> [31] "sumcov-1"  "sumcov-2"  "sumcov"    "sabscov-1" "sabscov-2"