Print a rgcca_bootstrap object

Print a fitted rgcca_bootstrap object. Parameters of the analysis and bootstrap statistics are displayed.

Usage

# S3 method for bootstrap
print(
  x,
  block = seq_along(x$rgcca$call$blocks),
  comp = 1,
  type = "weights",
  empirical = TRUE,
  display_order = FALSE,
  adj.method = "fdr",
  ...
)

Arguments

x

A fitted rgcca_bootstrap object (see rgcca_bootstrap)

block

A numeric corresponding to the block(s) to plot.

comp

A numeric vector indicating the component(s) to consider.

type

A string indicating the bootstrapped object to plot: block-weight vectors ("weights", default) or block-loading vectors ("loadings").

empirical

A logical value indicating if the bootstrap confidence intervals and p-values are derived from the empirical distribution. (default: TRUE)

display_order

A logical value for ordering the variables. If TRUE, variables are ordered from highest to lowest absolute value. If FALSE, the block order is used. Default is TRUE.

adj.method

A string indicating the method used to adjust the p-values. It must be a method handled by the p.adjust function. Default is "fdr".

...

Other parameters used in print (for the displaying of matrices).

Value

none

Examples

data("Russett")
blocks <- list(
  agriculture = Russett[, seq(3)],
  industry = Russett[, 4:5],
  politic = Russett[, 6:11]
)
fit.rgcca <- rgcca(blocks, ncomp = c(2, 1, 2))
boot.out <- rgcca_bootstrap(fit.rgcca, n_boot = 20, n_cores = 2)
#> Bootstrap samples sanity check...
#> OK
#>   |                                                  | 0 % ~calculating    |+++++                                             | 10% ~00s            |++++++++++                                        | 20% ~00s            |+++++++++++++++                                   | 30% ~00s            |++++++++++++++++++++                              | 40% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |+++++++++++++++++++++++++++++++++++               | 70% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
print(boot.out)
#> Call: method='rgcca', superblock=FALSE, scale=TRUE, scale_block='inertia',
#> init='svd', bias=TRUE, tol=1e-08, NA_method='na.ignore', ncomp=c(2,1,2),
#> response=NULL, comp_orth=TRUE 
#> There are J = 3 blocks.
#> The design matrix is:
#>             agriculture industry politic
#> agriculture           0        1       1
#> industry              1        0       1
#> politic               1        1       0
#> 
#> The factorial scheme is used.
#> 
#> Extracted statistics from 20 bootstrap samples.
#> Block-weight vectors for component 1: 
#>          estimate    mean     sd lower_bound upper_bound bootstrap_ratio   pval
#> gini       0.6476  0.6329 0.0646      0.4987       0.729          10.031 0.0000
#> farm       0.7603  0.7519 0.0562      0.6522       0.862          13.539 0.0000
#> rent      -0.0506 -0.0438 0.1631     -0.3462       0.232          -0.310 0.8182
#> gnpr       0.7044  0.7015 0.0428      0.6457       0.793          16.463 0.0000
#> labo      -0.7099 -0.7099 0.0489     -0.7636      -0.601         -14.509 0.0000
#> inst       0.1646  0.1664 0.0932     -0.0191       0.295           1.767 0.0526
#> ecks       0.4387  0.4340 0.0573      0.3450       0.523           7.653 0.0000
#> death      0.4581  0.4569 0.0406      0.3847       0.524          11.271 0.0000
#> demostab  -0.5647 -0.5538 0.0405     -0.6157      -0.473         -13.957 0.0000
#> demoinst   0.0362  0.0352 0.0941     -0.1540       0.186           0.385 0.3333
#> dictator   0.5004  0.4922 0.0473      0.4163       0.570          10.589 0.0000
#>          adjust.pval
#> gini          0.0000
#> farm          0.0000
#> rent          0.9091
#> gnpr          0.0000
#> labo          0.0000
#> inst          0.0957
#> ecks          0.0000
#> death         0.0000
#> demostab      0.0000
#> demoinst      0.4167
#> dictator      0.0000